feat: 장바구니 이동 모달을 컴포넌트로 분리 및 적용
- `CartConfirmModal` 컴포넌트를 생성하여 모달 내용을 별도 파일로 분리
- 부모 컴포넌트에서 user 상태를 props로 전달하도록 수정
- useEffect를 활용해 최초 렌더링 시 access token을 확인하여 user 상태 업데이트
- handleCloseModal을 props로 전달하여 모달 닫기 기능 적용
